+++ com.opendoorlogistics.core/src/com/opendoorlogistics/core/gis/map/tiled/TilePosition.java
@@ -85,7 +85,7 @@ class TilePosition {
 			}
 
 			@Override
-			public LatLong getLongLat(int pixelX, int pixelY) {
+			public LatLong getLongLat(double pixelX, double pixelY) {
 				throw new UnsupportedOperationException();
 			}
